Rework website landing/style-guide into a snapdesign showcase
Removes the old landing hero pieces (SvgLine, Landing, SnapSortPreview,
VisualDebuggerCard) and renames the style-guide component into a new
SnapDesign showcase, with /style-guide now redirecting to /snapdesign.
Also splits the SnapSort marketing page's inline hero/showcase markup out
into standalone components under routes/snapsort/components/, and adds
the framework logo icons used by the new showcase.
