+++ demo/svelte/main.ts
@@ -1,5 +1,5 @@
 import "../../css/snapdesign.scss";
-import "../../website/src/lib/landing/style-guide/style.dev.scss";
+import "../../website/src/lib/landing/snapdesign/style.dev.scss";
 import App from "./src/App.svelte";
 import { mount } from "svelte";
