--- website/src/lib/landing/Landing.svelte
@@ -1,13 +1,0 @@
-<script lang="ts">
-  import Hero from "./components/Hero.svelte";
-  import Highlights from "./components/Highlights.svelte";
-  import AssetsShowcase from "./components/AssetsShowcase.svelte";
-</script>
-
-<Hero />
-<Highlights />
-<AssetsShowcase />
-
-<style lang="scss">
-  @use "./landing.scss";
-</style>
