+++ website/src/routes/+page.svelte
@@ -1,5 +1,13 @@
 <script lang="ts">
-  import Landing from "$lib/landing/Landing.svelte";
+  import AssetsShowcase from "$lib/landing/components/AssetsShowcase.svelte";
+  import Hero from "$lib/landing/components/Hero.svelte";
+  import Highlights from "$lib/landing/components/Highlights.svelte";
 </script>
 
-<Landing />
+<Hero />
+<Highlights />
+<AssetsShowcase />
+
+<style lang="scss">
+  @use "$lib/landing/landing.scss";
+</style>
