+++ website/src/routes/snapdesign/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import SnapDesign from "$lib/landing/snapdesign/SnapDesign.svelte";
+</script>
+
+<SnapDesign />
