+++ website/src/routes/style-guide/+page.server.ts
@@ -1,0 +1,5 @@
+import { redirect } from "@sveltejs/kit";
+
+export function load() {
+  redirect(308, "/snapdesign");
+}
