+++ website/src/routes/style-guide/+page.svelte
@@ -1,5 +1,2 @@
 <script lang="ts">
-  import StyleGuide from "$lib/landing/style-guide/StyleGuide.svelte";
 </script>
-
-<StyleGuide />
